tests/infrastructure/requirements.txt
# Infrastructure testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
moto>=4.2.0
boto3>=1.34.0
aws-cdk-lib>=2.100.0
constructs>=10.3.0